Remove deprecation warning due to the SyntheticTestBuilder project put on hold
Problem: Currently the SyntheticTestBuilder is not completed, thus there is no alternative to the current approach for creating browser tests. This creates confusion since warnings are displayed when running tests, but there is no way to get rid of them.
Soltuion: Remove deprecation warning until SyntheticTestBuilder as an alternative solution is ready to use.